Move code into its own module, add docstrings
